projects:eval-adxl355-pmdz: Project folder structure re-organization
Re-organized src structure as follows:
├── common
│   ├── common_data.c
│   └── common_data.h
├── examples
│   ├── dummy
│   │   ├── dummy_example.c
│   │   └── dummy_example.h
│   ├── examples_src.mk
│   └── iio_example
│       ├── iio_example.c
│       └── iio_example.h
└── platform
    ├── platform_includes.h
    └── stm32
        ├── main.c
        ├── parameters.c
        ├── parameters.h
        └── platform_src.mk

This re-organization will make it easier to add new platform support
or to add new examples in the future.

Overview of the new structure:
common folder:
- contains common initialization data used by examples
- the data is platform independent
example folder:
- contains folders with examples
- contains makefile which includes the needed files based on the
selected examples
- all examples are platform independent
platform folder:
- contains folders with platforms
- contains makefile which includes the needed files based on the
selected platform
- contains main file which runs the selected example routines
