+++ projects/eval-adxl355-pmdz/Makefile
@@ -1,4 +1,6 @@
-TINYIIOD ?= y
+# Select the example you want to enable by chossing y for enabling and n for disabling
+DUMMY_EXAMPLE = n
+IIO_EXAMPLE = y
 
 include ../../tools/scripts/generic_variables.mk
