+++ projects/eval-adxl355-pmdz/src/examples/dummy/dummy_example.h
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   eval-adxl355-pmdz/src/parameters.h
- *   @brief  Parameters Definitions.
+ *   @file   dummy_example.h
+ *   @brief  DUMMY example header for eval-adxl355-pmdz project
  *   @author RBolboac (ramona.bolboaca@analog.com)
 ********************************************************************************
  * Copyright 2022(c) Analog Devices, Inc.
@@ -36,86 +36,16 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef __PARAMETERS_H__
-#define __PARAMETERS_H__
+#ifndef __DUMMY_EXAMPLE_H__
+#define __DUMMY_EXAMPLE_H__
 
 /******************************************************************************/
 /***************************** Include Files **********************************/
 /******************************************************************************/
 
-#ifdef XILINX_PLATFORM
-#include <xparameters.h>
-#endif
-
-#ifdef ADUCM_PLATFORM
-#include "irq_extra.h"
-#endif
-
-#ifdef STM32_PLATFORM
-#include "stm32_hal.h"
-#endif
-
 /******************************************************************************/
-/********************** Macros and Constants Definitions **********************/
+/************************ Functions Declarations ******************************/
 /******************************************************************************/
+int dummy_example_main ();
 
-
-#ifdef XILINX_PLATFORM
-
-#ifdef _XPARAMETERS_PS_H_
-#define ADC_DDR_BASEADDR	(XPAR_DDR_MEM_BASEADDR + 0x800000)
-#define DAC_DDR_BASEADDR	(XPAR_DDR_MEM_BASEADDR + 0xA000000)
-#define UART_DEVICE_ID		XPAR_XUARTPS_0_DEVICE_ID
-#define INTC_DEVICE_ID		XPAR_SCUGIC_SINGLE_DEVICE_ID
-
-#ifdef XPS_BOARD_ZCU102
-#define UART_IRQ_ID		XPAR_XUARTPS_0_INTR
-#else
-#define UART_IRQ_ID		XPAR_XUARTPS_1_INTR
-#endif
-
-#else // _XPARAMETERS_PS_H_
-
-#ifdef XPAR_DDR3_SDRAM_S_AXI_BASEADDR
-#define ADC_DDR_BASEADDR	(XPAR_DDR3_SDRAM_S_AXI_BASEADDR + 0x800000)
-#define DAC_DDR_BASEADDR	(XPAR_DDR3_SDRAM_S_AXI_BASEADDR + 0xA000000)
-#else
-#define ADC_DDR_BASEADDR	(XPAR_AXI_DDR_CNTRL_BASEADDR + 0x800000)
-#define DAC_DDR_BASEADDR	(XPAR_AXI_DDR_CNTRL_BASEADDR + 0xA000000)
-#endif
-
-#define UART_DEVICE_ID	XPAR_AXI_UART_DEVICE_ID
-#define INTC_DEVICE_ID	XPAR_INTC_SINGLE_DEVICE_ID
-#define UART_IRQ_ID		XPAR_AXI_INTC_AXI_UART_INTERRUPT_INTR
-#endif // _XPARAMETERS_PS_H_
-
-/* 400 * 8 * 2 = 6400 Default number of samples requested on a capture */
-#define MAX_SIZE_BASE_ADDR	10000
-#define UART_BAUDRATE	115200
-
-#endif // XILINX_PLATFORM
-
-#ifdef ADUCM_PLATFORM
-
-#define UART_DEVICE_ID	0
-#define INTC_DEVICE_ID	0
-#define UART_IRQ_ID		ADUCM_UART_INT_ID
-#define UART_BAUDRATE	115200
-
-#endif // ADUCM_PLATFORM
-
-#ifdef STM32_PLATFORM
-
-#define UART_DEVICE_ID	5
-extern UART_HandleTypeDef huart5;
-#define IIO_APP_HUART	(&huart5)
-#define UART_BAUDRATE	115200
-
-#endif // STM32_PLATFORM
-
-#ifdef USE_TCP_SOCKET
-#define WIFI_SSID	"RouterSSID"
-#define WIFI_PWD	"******"
-#endif /* USE_TCP_SOCKET */
-
-#endif // __PARAMETERS_H__
+#endif /* __DUMMY_EXAMPLE_H__ */
