+++ projects/eval-adxl355-pmdz/src/examples/examples_src.mk
@@ -1,0 +1,21 @@
+ifeq (y,$(strip $(IIO_EXAMPLE)))
+TINYIIOD=y
+CFLAGS += -DIIO_EXAMPLE
+SRCS += $(PROJECT)/src/examples/iio_example/iio_example.c
+INCS += $(PROJECT)/src/examples/iio_example/iio_example.h
+endif
+
+ifeq (y,$(strip $(DUMMY_EXAMPLE)))
+CFLAGS += -DDUMMY_EXAMPLE
+SRCS += $(PROJECT)/src/examples/dummy/dummy_example.c
+INCS += $(PROJECT)/src/examples/dummy/dummy_example.h
+endif
+
+ifeq (y,$(strip $(TINYIIOD)))
+SRC_DIRS += $(NO-OS)/iio/iio_app
+INCS += $(DRIVERS)/accel/adxl355/iio_adxl355.h
+SRCS += $(DRIVERS)/accel/adxl355/iio_adxl355.c
+
+INCS += $(INCLUDE)/no_os_list.h \
+		$(PLATFORM_DRIVERS)/uart_extra.h
+endif
